Move most of evaluation tests to ND4J
