+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/evaluation/EvaluationBinaryTest.java
@@ -14,13 +14,16 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package org.deeplearning4j.eval;
+package org.nd4j.evaluation;
 
-import org.deeplearning4j.BaseDL4JTest;
 import org.junit.Test;
+import org.nd4j.evaluation.classification.Evaluation;
+import org.nd4j.evaluation.classification.EvaluationBinary;
+import org.nd4j.linalg.BaseNd4jTest;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.api.ops.random.impl.BernoulliDistribution;
 import org.nd4j.linalg.factory.Nd4j;
+import org.nd4j.linalg.factory.Nd4jBackend;
 import org.nd4j.linalg.indexing.NDArrayIndex;
 
 import static org.junit.Assert.assertEquals;
@@ -28,7 +31,16 @@ import static org.junit.Assert.assertEquals;
 /**
  * Created by Alex on 20/03/2017.
  */
-public class EvaluationBinaryTest extends BaseDL4JTest {
+public class EvaluationBinaryTest extends BaseNd4jTest {
+
+    public EvaluationBinaryTest(Nd4jBackend backend) {
+        super(backend);
+    }
+
+    @Override
+    public char ordering() {
+        return 'c';
+    }
 
     @Test
     public void testEvaluationBinary() {
