+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/evaluation/ROCBinaryTest.java
@@ -14,21 +14,33 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package org.deeplearning4j.eval;
+package org.nd4j.evaluation;
 
-import org.deeplearning4j.BaseDL4JTest;
 import org.junit.Test;
+import org.nd4j.evaluation.classification.ROC;
+import org.nd4j.evaluation.classification.ROCBinary;
 import org.nd4j.evaluation.curves.PrecisionRecallCurve;
+import org.nd4j.linalg.BaseNd4jTest;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.api.ops.random.impl.BernoulliDistribution;
 import org.nd4j.linalg.factory.Nd4j;
+import org.nd4j.linalg.factory.Nd4jBackend;
 
 import static org.junit.Assert.assertEquals;
 
 /**
  * Created by Alex on 21/03/2017.
  */
-public class ROCBinaryTest extends BaseDL4JTest {
+public class ROCBinaryTest extends BaseNd4jTest {
+
+    public ROCBinaryTest(Nd4jBackend backend) {
+        super(backend);
+    }
+
+    @Override
+    public char ordering() {
+        return 'c';
+    }
 
     @Test
     public void testROCBinary() {
